Separate component from `formRow.tsx` to multi files
